Update app.js
In reality I would want to hide my API key in an env file but I can't do that using github pages which requires all files in the repository to be public. My key is locked so it can't be abused so until I move away from github pages I'm going to leave it as is.